+++ deploy/kubernetes/helm/sloth/tests/values_test.go
@@ -8,12 +8,17 @@ func defaultValues() msi {
 
 func customValues() msi {
 	return msi{
+		"global": msi{
+			"imageRegistry": "",
+		},
+
 		"labels": msi{
 			"label-from": "test",
 		},
 
 		"image": msi{
-			"repository": "slok/sloth-test",
+			"registry":   "slok",
+			"repository": "sloth-test",
 			"tag":        "v1.42.42",
 		},